refactor: remove duplicate code from competition
